requirements.txt
aiohttp==3.8.1
python-telegram-bot==20.0
requests==2.26.0
pandas==1.3.3
numpy==1.21.2
loguru==0.5.3